refactoring and deploy on docker
