+++ api_license/.dockerignore
@@ -1,0 +1,1 @@
+env_api_license
