+++ api_license/app/api/endpoints/process_document.py
@@ -1,0 +1,35 @@
+from fastapi import APIRouter, File, UploadFile, Form  
+from app.services.ocr import extract_text_from_image  
+from app.services.nlp import person_name_entity_recognition  
+from app.services.utils import clean_and_split_text, find_bounding_boxes, fuzzy_match_names  
+import pytesseract  
+import cv2  
+import numpy as np  
+  
+router = APIRouter()  
+  
+@router.post("/")  
+async def process_document(file: UploadFile = File(...), name: str = Form(...)):  
+    contents = await file.read()  
+    nparr = np.frombuffer(contents, np.uint8)  
+    image = cv2.imdecode(nparr, cv2.IMREAD_COLOR)  
+      
+    extracted_text, processed_img = extract_text_from_image(image)  
+    cleaned_text = clean_and_split_text(extracted_text)  
+    person_names = person_name_entity_recognition(cleaned_text)  
+    ocr_data = pytesseract.image_to_data(processed_img, output_type=pytesseract.Output.DICT)  
+      
+    results = []  
+    bounding_boxes = find_bounding_boxes(person_names[0].split(), ocr_data)  
+  
+    print(f"Nombre proporcionado: {name}")  
+    print(f"Nombres reconocidos: {person_names}")  
+  
+    fuzzy_similarity, similar_name = fuzzy_match_names(person_names[0], name)  
+    results.append({  
+        "extracted_name": person_names[0],  
+        "bounding_box": bounding_boxes,  
+        "fuzzy_similarity": fuzzy_similarity,  
+        "similar_name": similar_name  
+    })  
+    return {"results": results}
