+++ api_license/run.sh
@@ -1,0 +1,4 @@
+#!/bin/sh  
+  
+# Run the app  
+uvicorn app.main:app --host 0.0.0.0 --port 8000
